tests/isaacsim/streaming/requirements.txt
fastapi==0.68.1
uvicorn==0.15.0
httpx==0.23.0
pyjwt==2.3.0 